refactor(webpack): move webpack configuration to only two files
