+++ packages/arui-scripts/src/configs/webpack.client.dev.ts
@@ -1,307 +1,9 @@
-import path from 'path';
-import webpack from 'webpack';
-import CaseSensitivePathsPlugin from 'case-sensitive-paths-webpack-plugin';
-import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-import AssetsPlugin from 'assets-webpack-plugin';
-import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
-import getCSSModuleLocalIdent from 'react-dev-utils/getCSSModuleLocalIdent';
-const PnpWebpackPlugin = require('pnp-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-const ReactRefreshTypeScript = require('react-refresh-typescript');
-
-import configs from './app-configs';
-import babelConf from './babel-client';
-import { babelDependencies } from './babel-dependencies';
-import postcssConf from './postcss';
 import applyOverrides from './util/apply-overrides';
-import getEntry from './util/get-entry';
+import { createClientWebpackConfig } from './webpack.client';
 
-// style files regexes
-const cssRegex = /\.css$/;
-const cssModuleRegex = /\.module\.css$/;
+const config = applyOverrides(
+    ['webpack', 'webpackClient', 'webpackDev', 'webpackClientDev'],
+    createClientWebpackConfig('dev'),
+);
 
-function getSingleEntry(clientEntry: string[]) {
-    return [
-        ...(Array.isArray(configs.clientPolyfillsEntry) ? configs.clientPolyfillsEntry : [configs.clientPolyfillsEntry]),
-        require.resolve('webpack/hot/dev-server'),
-        // Finally, this is your app's code:
-        ...clientEntry,
-        // We include the app code last so that if there is a runtime error during
-        // initialization, it doesn't blow up the WebpackDevServer client, and
-        // changing JS code would still trigger a refresh.
-    ].filter(Boolean) as string[];
-}
-
-// This is the development configuration.
-// It is focused on developer experience and fast rebuilds.
-// The production configuration is different and lives in a separate file.
-const webpackClientDev = applyOverrides(['webpack', 'webpackClient', 'webpackDev', 'webpackClientDev'], {
-    target: 'web',
-    mode: 'development',
-    devtool: configs.devSourceMaps,
-    // These are the "entry points" to our application.
-    // This means they will be the "root" imports that are included in JS bundle.
-    entry: getEntry(configs.clientEntry, getSingleEntry),
-    context: configs.cwd,
-    output: {
-        // Add /* filename */ comments to generated require()s in the output.
-        pathinfo: true,
-        path: configs.clientOutputPath,
-        publicPath: configs.publicPath,
-        filename: '[name].js',
-        chunkFilename: '[name].js',
-        // Point sourcemap entries to original disk location (format as URL on Windows)
-        devtoolModuleFilenameTemplate: (info: any) =>
-            path
-                .relative(configs.appSrc, info.absoluteResourcePath)
-                .replace(/\\/g, '/'),
-    },
-    resolve: {
-        // This allows you to set a fallback for where Webpack should look for modules.
-        // We placed these paths second because we want `node_modules` to "win"
-        // if there are any conflicts. This matches Node resolution mechanism.
-        // https://github.com/facebookincubator/create-react-app/issues/253
-        modules: ['node_modules', configs.appNodeModules],
-        // These are the reasonable defaults supported by the Node ecosystem.
-        // We also include JSX as a common component filename extension to support
-        // some tools, although we do not recommend using it, see:
-        // https://github.com/facebookincubator/create-react-app/issues/290
-        // `web` extension prefixes have been added for better support
-        // for React Native Web.
-        extensions: ['.web.js', '.mjs', '.js', '.json', '.web.jsx', '.jsx', '.ts', '.tsx'],
-        plugins: ([
-            (configs.tsconfig && new TsconfigPathsPlugin({
-                configFile: configs.tsconfig,
-                extensions: ['.web.js', '.mjs', '.js', '.json', '.web.jsx', '.jsx', '.ts', '.tsx']
-            }))
-        ].filter(Boolean)) as NonNullable<webpack.Configuration['resolve']>['plugins'],
-    },
-    resolveLoader: {
-        plugins: [
-            PnpWebpackPlugin.moduleLoader(module),
-        ],
-    },
-    cache: {
-        type: 'filesystem',
-        buildDependencies: {
-            config: [__filename],
-        },
-    },
-    module: {
-        // typescript interface will be removed from modules, and we will get an error on correct code
-        // see https://github.com/webpack/webpack/issues/7378
-        strictExportPresence: !configs.tsconfig,
-        rules: [
-            {
-                // "oneOf" will traverse all following loaders until one will
-                // match the requirements. When no loader matches it will fall
-                // back to the "file" loader at the end of the loader list.
-                oneOf: ([
-                    {
-                        test: [/\.svg$/],
-                        loader: require.resolve('svg-url-loader'),
-                        options: {
-                            limit: 10000,
-                            iesafe: true,
-                            name: '[name].[hash:8].[ext]',
-                        },
-                    },
-                    // "url" loader works like "file" loader except that it embeds assets
-                    // smaller than specified limit in bytes as data URLs to avoid requests.
-                    // A missing `test` is equivalent to a match.
-                    {
-                        test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
-                        loader: require.resolve('url-loader'),
-                        options: {
-                            limit: 10000,
-                            name: '[name].[hash:8].[ext]',
-                        },
-                    },
-                    // Process JS with Babel.
-                    {
-                        test: configs.useTscLoader ? /\.(js|jsx|mjs|cjs)$/ : /\.(js|jsx|mjs|ts|tsx|cjs)$/,
-                        include: configs.appSrc,
-                        use: [
-                            {
-                                loader: require.resolve('babel-loader'),
-                                options: Object.assign({
-                                    // This is a feature of `babel-loader` for webpack (not Babel itself).
-                                    // It enables caching results in ./node_modules/.cache/babel-loader/
-                                    // directory for faster rebuilds.
-                                    cacheDirectory: true,
-                                    plugins: require.resolve('react-refresh/babel')
-                                }, babelConf)
-                            }
-                        ]
-                    },
-                    (configs.tsconfig && configs.useTscLoader) && {
-                        test: /\.tsx?$/,
-                        use: [
-                            {
-                                loader: require.resolve('babel-loader'),
-                                options: Object.assign({
-                                    // This is a feature of `babel-loader` for webpack (not Babel itself).
-                                    // It enables caching results in ./node_modules/.cache/babel-loader/
-                                    // directory for faster rebuilds.
-                                    cacheDirectory: true,
-                                    cacheCompression: false,
-                                    plugins: require.resolve('react-refresh/babel')
-                                }, babelConf)
-                            },
-                            {
-                                loader: require.resolve('ts-loader'),
-                                options: {
-                                    getCustomTransformers: () => ({
-                                        before: [ReactRefreshTypeScript()],
-                                    }),
-                                    onlyCompileBundledFiles: true,
-                                    transpileOnly: true,
-                                    happyPackMode: true,
-                                    configFile: configs.tsconfig
-                                }
-                            }
-                        ]
-                    },
-                    // Process any JS outside of the app with Babel.
-                    // Unlike the application JS, we only compile the standard ES features.
-                    {
-                        test: /\.(js|mjs)$/,
-                        exclude: /@babel(?:\/|\\{1,2})runtime/,
-                        loader: require.resolve('babel-loader'),
-                        options: {
-                            ...babelDependencies,
-                            babelrc: false,
-                            configFile: false,
-                            compact: false,
-                            cacheDirectory: true,
-                            cacheCompression: false,
-                        },
-                    },
-                    // "postcss" loader applies autoprefixer to our CSS.
-                    // "css" loader resolves paths in CSS and adds assets as dependencies.
-                    // "style" loader turns CSS into JS modules that inject <style> tags.
-                    // In production, we use a plugin to extract that CSS to a file, but
-                    // in development "style" loader enables hot editing of CSS.
-                    {
-                        test: cssRegex,
-                        exclude: cssModuleRegex,
-                        use: [
-                            {
-                                loader: MiniCssExtractPlugin.loader,
-                                options: {
-                                    publicPath: './',
-                                },
-                            },
-                            {
-                                loader: require.resolve('css-loader'),
-                                options: {
-                                    importLoaders: 1,
-                                },
-                            },
-                            {
-                                loader: require.resolve('postcss-loader'),
-                                options: {
-                                    // Necessary for external CSS imports to work
-                                    // https://github.com/facebookincubator/create-react-app/issues/2677
-                                    ident: 'postcss',
-                                    plugins: () => postcssConf,
-                                },
-                            },
-                        ],
-                    },
-                    {
-                        test: cssModuleRegex,
-                        use: [
-                            {
-                                loader: MiniCssExtractPlugin.loader,
-                                options: {
-                                    publicPath: './',
-                                },
-                            },
-                            {
-                                loader: require.resolve('css-loader'),
-                                options: {
-                                    importLoaders: 1,
-                                    modules: {
-                                        getLocalIdent: getCSSModuleLocalIdent
-                                    },
-                                },
-                            },
-                            {
-                                loader: require.resolve('postcss-loader'),
-                                options: {
-                                    // Necessary for external CSS imports to work
-                                    // https://github.com/facebookincubator/create-react-app/issues/2677
-                                    ident: 'postcss',
-                                    plugins: () => postcssConf,
-                                },
-                            },
-                        ],
-                    },
-                    // "file" loader makes sure those assets get served by WebpackDevServer.
-                    // When you `import` an asset, you get its (virtual) filename.
-                    // In production, they would get copied to the `build` folder.
-                    // This loader doesn't use a "test" so it will catch all modules
-                    // that fall through the other loaders.
-                    {
-                        // Exclude `js` files to keep "css" loader working as it injects
-                        // its runtime that would otherwise processed through "file" loader.
-                        // Also exclude `html` and `json` extensions so they get processed
-                        // by webpacks internal loaders.
-                        exclude: [/\.(js|jsx|mjs|ts|tsx)$/, /\.(html|ejs)$/, /\.json$/],
-                        loader: require.resolve('file-loader'),
-                        options: {
-                            name: 'static/media/[name].[hash:8].[ext]',
-                        },
-                    },
-                ].filter(Boolean)) as webpack.RuleSetRule[],
-            },
-            // ** STOP ** Are you adding a new loader?
-            // Make sure to add the new loader(s) before the "file" loader.
-        ],
-    },
-    plugins: ([
-        new AssetsPlugin({ path: configs.serverOutputPath }),
-        new webpack.DefinePlugin({
-            'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'development'),
-            // Tell Webpack to provide empty mocks for process.env.
-            'process.env': '{}'
-        }),
-        new ReactRefreshWebpackPlugin({
-            overlay: {
-                sockIntegration: 'whm',
-            },
-        }),
-        // Watcher doesn't work well if you mistype casing in a path so we use
-        // a plugin that prints an error when you attempt to do this.
-        // See https://github.com/facebookincubator/create-react-app/issues/240
-        new CaseSensitivePathsPlugin(),
-        configs.tsconfig !== null && new ForkTsCheckerWebpackPlugin(),
-        new MiniCssExtractPlugin(),
-    ].filter(Boolean)),
-    // Turn off performance hints during development because we don't do any
-    // splitting or minification in interest of speed. These warnings become
-    // cumbersome.
-    performance: {
-        hints: false,
-    },
-    optimization: {
-        minimize: false,
-        nodeEnv: false,
-        // Оптимизации времени билда, см https://webpack.js.org/guides/build-performance/
-        removeAvailableModules: false,
-        removeEmptyChunks: false,
-        splitChunks: false
-    },
-    experiments: {
-        backCompat: configs.webpack4Compatibility,
-    },
-    // Без этого комиляция трирегилась на изменение в node_modules и приводила к утечке памяти
-    watchOptions: {
-        ignored: new RegExp(configs.watchIgnorePath.join('|')),
-        aggregateTimeout: 100,
-    }
-});
-
-export default webpackClientDev;
+export default config;
